Reorganize fetchcache to use index - Update all imports
